Delete PrivilegeChecker.java unused code
