+++ fe/fe-core/src/main/java/com/starrocks/sql/analyzer/PrivilegeStmtAnalyzer.java
@@ -16,20 +16,29 @@ package com.starrocks.sql.analyzer;
 
 import com.google.common.base.Preconditions;
 import com.google.common.base.Strings;
-import com.starrocks.analysis.ResourcePattern;
-import com.starrocks.analysis.TablePattern;
-import com.starrocks.catalog.AccessPrivilege;
+import com.google.common.collect.Lists;
+import com.starrocks.analysis.FunctionName;
+import com.starrocks.analysis.TableName;
+import com.starrocks.authentication.AuthenticationException;
+import com.starrocks.authentication.AuthenticationManager;
+import com.starrocks.authentication.AuthenticationProvider;
+import com.starrocks.authentication.AuthenticationProviderFactory;
+import com.starrocks.authentication.PlainPasswordAuthenticationProvider;
+import com.starrocks.authentication.UserAuthenticationInfo;
+import com.starrocks.catalog.Database;
+import com.starrocks.catalog.Function;
+import com.starrocks.catalog.FunctionSearchDesc;
 import com.starrocks.common.AnalysisException;
 import com.starrocks.common.Config;
 import com.starrocks.common.DdlException;
-import com.starrocks.common.ErrorCode;
-import com.starrocks.common.ErrorReport;
+import com.starrocks.common.Pair;
 import com.starrocks.mysql.MysqlPassword;
-import com.starrocks.mysql.privilege.Auth;
-import com.starrocks.mysql.privilege.AuthPlugin;
-import com.starrocks.mysql.privilege.PrivBitSet;
-import com.starrocks.mysql.privilege.Privilege;
-import com.starrocks.mysql.privilege.Role;
+import com.starrocks.privilege.AuthorizationManager;
+import com.starrocks.privilege.ObjectType;
+import com.starrocks.privilege.PEntryObject;
+import com.starrocks.privilege.PrivilegeBuiltinConstants;
+import com.starrocks.privilege.PrivilegeException;
+import com.starrocks.privilege.PrivilegeType;
 import com.starrocks.qe.ConnectContext;
 import com.starrocks.server.GlobalStateMgr;
 import com.starrocks.sql.ast.AlterUserStmt;
@@ -38,300 +47,564 @@ import com.starrocks.sql.ast.BaseCreateAlterUserStmt;
 import com.starrocks.sql.ast.BaseGrantRevokePrivilegeStmt;
 import com.starrocks.sql.ast.BaseGrantRevokeRoleStmt;
 import com.starrocks.sql.ast.CreateRoleStmt;
+import com.starrocks.sql.ast.CreateUserStmt;
 import com.starrocks.sql.ast.DropRoleStmt;
 import com.starrocks.sql.ast.DropUserStmt;
 import com.starrocks.sql.ast.ExecuteAsStmt;
-import com.starrocks.sql.ast.GrantPrivilegeStmt;
+import com.starrocks.sql.ast.FunctionArgsDef;
+import com.starrocks.sql.ast.SetDefaultRoleStmt;
+import com.starrocks.sql.ast.SetRoleStmt;
+import com.starrocks.sql.ast.ShowAuthenticationStmt;
 import com.starrocks.sql.ast.ShowGrantsStmt;
 import com.starrocks.sql.ast.StatementBase;
 import com.starrocks.sql.ast.UserIdentity;
-import org.jetbrains.annotations.NotNull;
+import com.starrocks.sql.common.MetaUtils;
+import org.apache.commons.lang3.EnumUtils;
 
+import java.nio.charset.StandardCharsets;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
+import java.util.Set;
 
 public class PrivilegeStmtAnalyzer {
+    private PrivilegeStmtAnalyzer() {
+    }
+
     public static void analyze(StatementBase statement, ConnectContext session) {
-        if (session.getGlobalStateMgr().isUsingNewPrivilege()) {
-            PrivilegeStmtAnalyzerV2.analyze(statement, session);
-        } else {
-            new PrivilegeStatementAnalyzerVisitor().analyze(statement, session);
-        }
+        new PrivilegeStatementAnalyzerVisitor().analyze(statement, session);
     }
 
     static class PrivilegeStatementAnalyzerVisitor extends AstVisitor<Void, ConnectContext> {
+        private AuthenticationManager authenticationManager = null;
+        private AuthorizationManager authorizationManager = null;
+
         public void analyze(StatementBase statement, ConnectContext session) {
+            authenticationManager = GlobalStateMgr.getCurrentState().getAuthenticationManager();
+            authorizationManager = GlobalStateMgr.getCurrentState().getAuthorizationManager();
             visit(statement, session);
         }
 
         /**
          * analyse user identity + check if user exists in UserPrivTable
          */
-        private void analyseUser(UserIdentity userIdent, ConnectContext session, boolean checkExist) {
-            // analyse user identity
+        private void analyseUser(UserIdentity userIdent, boolean checkExist) {
             userIdent.analyze();
 
-            if (checkExist) {
-                // check if user exists
-                if (!session.getGlobalStateMgr().getAuth().doesUserExist(userIdent)) {
-                    throw new SemanticException("user " + userIdent + " not exist!");
-                }
+            // check if user exists
+            if (checkExist && !authenticationManager.doesUserExist(userIdent)) {
+                throw new SemanticException("cannot find user " + userIdent + "!");
             }
         }
 
         /**
-         * check if role name valid and get full role name + check if role exists
+         * check if role name valid and get full role name
          */
-        private void analyseRoleName(String roleName, ConnectContext session, boolean canBeAdmin, String errMsg) {
-            validRoleName(roleName, canBeAdmin, errMsg);
-            if (!session.getGlobalStateMgr().getAuth().doesRoleExist(roleName)) {
-                throw new SemanticException("role " + roleName + " not exist!");
+        private void validRoleName(String roleName, String errMsg, boolean checkExist) {
+            // always set to true, we can validate if it's allowed to operation on admin later
+            FeNameFormat.checkRoleName(roleName, true, errMsg);
+            // check if role exists
+            if (checkExist && !authorizationManager.checkRoleExists(roleName)) {
+                throw new SemanticException(errMsg + ": cannot find role " + roleName + "!");
             }
         }
 
         /**
-         * check if role name valid
+         * Get scrambled password from plain password
          */
-        private void validRoleName(String roleName, boolean canBeAdmin, String errMsg) {
-            FeNameFormat.checkRoleName(roleName, canBeAdmin, errMsg);
-        }
-
-        /**
-         * GRANT rolexx to userxx
-         * REVOKE rolexx from userxx
-         */
-        @Override
-        public Void visitGrantRevokeRoleStatement(BaseGrantRevokeRoleStmt stmt, ConnectContext session) {
-            if (stmt.getUserIdentity() == null) {
-                throw new SemanticException("Unsupported syntax: grant/revoke to role is not supported");
-            }
-            analyseUser(stmt.getUserIdentity(), session, true);
-            stmt.getGranteeRole().forEach(role -> analyseRoleName(role, session, true, "Can not granted/revoke role to user"));
-            return null;
-        }
-
-        @Override
-        public Void visitGrantRevokePrivilegeStatement(BaseGrantRevokePrivilegeStmt stmt, ConnectContext session) {
-            if (stmt instanceof GrantPrivilegeStmt && ((GrantPrivilegeStmt) stmt).isWithGrantOption()) {
-                throw new SemanticException("unsupported syntax: WITH GRANT OPTION");
-            }
-            // validate user/role
-            if (stmt.getUserIdentity() != null) {
-                analyseUser(stmt.getUserIdentity(), session, true);
-            } else {
-                analyseRoleName(stmt.getRole(), session, true, "invalid role");
-            }
-
-            // parse privilege actions to PrivBitSet
-            PrivBitSet privs = getPrivBitSet(stmt.getPrivilegeTypeUnResolved());
-            String privType = stmt.getObjectTypeUnResolved();
-            if (privType.equals("TABLE") || privType.equals("DATABASE")) {
-                if (stmt.getPrivilegeObjectNameTokensList().size() != 1) {
-                    throw new SemanticException("unsupported syntax: can only grant/revoke on one " + privType);
-                }
-                analyseTablePrivs(stmt, privs, stmt.getPrivilegeObjectNameTokensList().get(0));
-            } else if (privType.equals("RESOURCE")) {
-                if (stmt.getPrivilegeObjectNameTokensList().size() != 1) {
-                    throw new SemanticException("unsupported syntax: can only grant/revoke on one " + privType);
-                }
-                analyseResourcePrivs(stmt, privs, stmt.getPrivilegeObjectNameTokensList().get(0));
-            } else if (privType.equals("USER")) {
-                if (stmt.getPrivilegeTypeUnResolved().size() != 1 || !privs.containsPrivs(Privilege.IMPERSONATE_PRIV)) {
-                    throw new SemanticException("only IMPERSONATE can only be granted on user");
-                }
-                if (stmt.getUserPrivilegeObjectList().size() != 1) {
-                    throw new SemanticException("unsupported syntax: can only grant/revoke on one USER");
-                }
-                stmt.setPrivBitSet(privs);
-                analyseUser(stmt.getUserPrivilegeObjectList().get(0), session, true);
-            } else {
-                throw new SemanticException("unsupported privilege type " + privType);
-            }
-
-            return null;
-        }
-
-        @NotNull
-        private static PrivBitSet getPrivBitSet(List<String> privList) {
-            if (privList.isEmpty()) {
-                throw new SemanticException("No privileges in grant statement.");
-            }
-            PrivBitSet privs = PrivBitSet.of();
-            for (String privStr : privList) {
-                AccessPrivilege accessPrivilege;
-                try {
-                    accessPrivilege = AccessPrivilege.valueOf(privStr);
-                } catch (IllegalArgumentException e) {
-                    try {
-                        // SELECT -> SELECT_PRIV
-                        accessPrivilege = AccessPrivilege.valueOf(privStr + "_PRIV");
-                    } catch (IllegalArgumentException e1) {
-                        throw new SemanticException("Unknown privilege " + privStr);
-                    }
-                }
-                if (accessPrivilege == null) {
-                    throw new SemanticException("Unknown privilege " + privStr);
-                }
-                privs.or(accessPrivilege.toPrivilege());
-            }
-            return privs;
-        }
-
-        private static void analyseResourcePrivs(BaseGrantRevokePrivilegeStmt stmt, PrivBitSet privs,
-                                                 List<String> privilegeObjectList) {
-            if (privilegeObjectList == null || privilegeObjectList.size() != 1) {
-                throw new SemanticException("invalid resource pattern!");
-            }
-            ResourcePattern resourcePattern = new ResourcePattern(privilegeObjectList.get(0));
-            try {
-                resourcePattern.analyze();
-            } catch (AnalysisException e) {
-                SemanticException exception = new SemanticException("invalid resource pattern " + resourcePattern);
-                exception.initCause(e);
-                throw exception;
-            }
-            if (resourcePattern.getPrivLevel() != Auth.PrivLevel.GLOBAL
-                    && privs.containsPrivs(Privilege.NODE_PRIV, Privilege.ADMIN_PRIV)) {
-                throw new SemanticException(privs.toPrivilegeList() + " can only be granted on resource *");
-            }
-            stmt.setAnalysedResource(privs, resourcePattern);
-        }
-
-        private static void analyseTablePrivs(BaseGrantRevokePrivilegeStmt stmt, PrivBitSet privs,
-                                              List<String> privilegeObjectList) {
-            if (privilegeObjectList == null) {
-                throw new SemanticException("invalid table pattern!");
-            }
-            TablePattern tablePattern;
-            if (privilegeObjectList.size() == 1) {
-                tablePattern = new TablePattern(privilegeObjectList.get(0), "*");
-            } else {
-                tablePattern = new TablePattern(privilegeObjectList.get(0), privilegeObjectList.get(1));
-            }
-            try {
-                tablePattern.analyze();
-            } catch (AnalysisException e) {
-                SemanticException exception = new SemanticException("invalid table pattern " + tablePattern.toString());
-                exception.initCause(e);
-                throw exception;
-            }
-            if (tablePattern.getPrivLevel() != Auth.PrivLevel.GLOBAL
-                    && privs.containsPrivs(Privilege.NODE_PRIV, Privilege.ADMIN_PRIV)) {
-                throw new SemanticException(privs.toPrivilegeList() + " can only be granted on table *.*");
-            }
-            stmt.setAnalysedTable(privs, tablePattern);
-        }
-
-        /**
-         * analyse EXECUTE AS statment
-         */
-        @Override
-        public Void visitExecuteAsStatement(ExecuteAsStmt stmt, ConnectContext session) {
-            if (stmt.isAllowRevert()) {
-                throw new SemanticException("`EXECUTE AS` must use with `WITH NO REVERT` for now!");
-            }
-            analyseUser(stmt.getToUser(), session, true);
-            return null;
-        }
-
-        /**
-         * convert password to hashed password
-         */
-        private byte[] analysePassword(UserIdentity userIdent, String originalPassword, boolean isPasswordPlain,
-                                       boolean checkReuse) {
+        private byte[] analysePassword(String originalPassword, boolean isPasswordPlain) {
             if (Strings.isNullOrEmpty(originalPassword)) {
-                return new byte[0];
+                return MysqlPassword.EMPTY_PASSWORD;
             }
             try {
                 if (isPasswordPlain) {
-                    // plain password should check for validation & reuse
-                    Auth.validatePassword(originalPassword);
-                    if (checkReuse) {
-                        GlobalStateMgr.getCurrentState().getAuth().checkPasswordReuse(userIdent, originalPassword);
-                    }
-                    // convert plain password to scramble
                     return MysqlPassword.makeScrambledPassword(originalPassword);
                 } else {
                     return MysqlPassword.checkPassword(originalPassword);
                 }
-            } catch (DdlException | AnalysisException e) {
-                // TODO AnalysisException used to raise in all old methods is captured and translated to SemanticException
-                // that is permitted to throw during analyzing phrase under the new framework for compatibility.
-                // Remove it after all old methods migrate to the new framework
+            } catch (AnalysisException e) {
                 throw new SemanticException(e.getMessage());
             }
         }
 
+        @Override
         public Void visitBaseCreateAlterUserStmt(BaseCreateAlterUserStmt stmt, ConnectContext session) {
-            analyseUser(stmt.getUserIdentity(), session, stmt instanceof AlterUserStmt);
-            /*
-             * IDENTIFIED BY
-             */
-            stmt.setScramblePassword(
-                    analysePassword(stmt.getUserIdentity(), stmt.getOriginalPassword(), stmt.isPasswordPlain(),
-                            stmt instanceof AlterUserStmt));
-            /*
-             * IDENTIFIED WITH
-             */
-            if (!Strings.isNullOrEmpty(stmt.getAuthPluginName())) {
-                if (AuthPlugin.AUTHENTICATION_LDAP_SIMPLE.name().equals(stmt.getAuthPluginName())) {
-                    stmt.setUserForAuthPlugin(stmt.getAuthStringUnResolved());
-                } else if (AuthPlugin.MYSQL_NATIVE_PASSWORD.name().equals(stmt.getAuthPluginName())) {
-                    // in this case, authString is password
-                    stmt.setScramblePassword(analysePassword(stmt.getUserIdentity(), stmt.getAuthStringUnResolved(),
-                            stmt.isPasswordPlain(), stmt instanceof AlterUserStmt));
-                } else if (AuthPlugin.AUTHENTICATION_KERBEROS.name().equalsIgnoreCase(stmt.getAuthPluginName()) &&
-                        GlobalStateMgr.getCurrentState().getAuth().isSupportKerberosAuth()) {
-                    // In kerberos authentication, userForAuthPlugin represents the user principal realm.
-                    // If user realm is not specified when creating user, the service principal realm will be used as
-                    // the user principal realm by default.
-                    if (stmt.getAuthStringUnResolved() != null) {
-                        stmt.setUserForAuthPlugin(stmt.getAuthStringUnResolved());
-                    } else {
-                        stmt.setUserForAuthPlugin(Config.authentication_kerberos_service_principal.split("@")[1]);
-                    }
-                } else {
-                    ErrorReport.reportSemanticException(ErrorCode.ERR_AUTH_PLUGIN_NOT_LOADED, stmt.getAuthPluginName());
+            stmt.getUserIdentity().analyze();
+            if (stmt instanceof CreateUserStmt) {
+                CreateUserStmt createUserStmt = (CreateUserStmt) stmt;
+                if (authenticationManager.doesUserExist(createUserStmt.getUserIdentity()) && !createUserStmt.isIfNotExists()) {
+                    throw new SemanticException("Operation CREATE USER failed for " + createUserStmt.getUserIdentity()
+                            + " : user already exists");
+                }
+            } else {
+                AlterUserStmt alterUserStmt = (AlterUserStmt) stmt;
+                if (!authenticationManager.doesUserExist(stmt.getUserIdentity()) && !alterUserStmt.isIfExists()) {
+                    throw new SemanticException("Operation ALTER USER failed for " + alterUserStmt.getUserIdentity()
+                            + " : user not exists");
                 }
             }
 
-            if (!stmt.getDefaultRoles().isEmpty()) {
-                String role = stmt.getDefaultRoles().get(0);
-
-                if (role.equalsIgnoreCase("SUPERUSER")) {
-                    // for forward compatibility
-                    stmt.getDefaultRoles().set(0, Role.ADMIN_ROLE);
+            byte[] password = MysqlPassword.EMPTY_PASSWORD;
+            String authPluginUsing;
+            if (stmt.getAuthPluginName() == null) {
+                authPluginUsing = authenticationManager.getDefaultPlugin();
+                password = analysePassword(stmt.getOriginalPassword(), stmt.isPasswordPlain());
+            } else {
+                authPluginUsing = stmt.getAuthPluginName();
+                if (authPluginUsing.equals(PlainPasswordAuthenticationProvider.PLUGIN_NAME)) {
+                    // In this case, authString is the password
+                    password = analysePassword(stmt.getAuthStringUnResolved(), stmt.isPasswordPlain());
                 }
-                analyseRoleName(role, session, true, "Can not granted/revoke role to user");
+            }
+
+            try {
+                AuthenticationProvider provider = AuthenticationProviderFactory.create(authPluginUsing);
+                UserIdentity userIdentity = stmt.getUserIdentity();
+                UserAuthenticationInfo info = provider.validAuthenticationInfo(
+                        userIdentity, new String(password, StandardCharsets.UTF_8), stmt.getAuthStringUnResolved());
+                info.setAuthPlugin(authPluginUsing);
+                info.setOrigUserHost(userIdentity.getQualifiedUser(), userIdentity.getHost());
+                stmt.setAuthenticationInfo(info);
+                if (stmt instanceof AlterUserStmt) {
+                    session.getGlobalStateMgr().getAuthenticationManager().checkPasswordReuse(
+                            userIdentity, stmt.getOriginalPassword());
+                }
+            } catch (AuthenticationException | DdlException e) {
+                SemanticException exception = new SemanticException("invalidate authentication: " + e.getMessage());
+                exception.initCause(e);
+                throw exception;
+            }
+
+            if (!stmt.getDefaultRoles().isEmpty()) {
+                stmt.getDefaultRoles().forEach(r -> validRoleName(r, "Valid role name fail", true));
+            }
+            return null;
+        }
+
+        private boolean needProtectAdminUser(UserIdentity userIdentity, ConnectContext context) {
+            return Config.authorization_enable_admin_user_protection &&
+                    userIdentity.getQualifiedUser().equalsIgnoreCase("admin") &&
+                    !context.getCurrentUserIdentity().equals(UserIdentity.ROOT);
+        }
+
+        @Override
+        public Void visitDropUserStatement(DropUserStmt stmt, ConnectContext session) {
+            UserIdentity userIdentity = stmt.getUserIdentity();
+            userIdentity.analyze();
+
+            if (needProtectAdminUser(userIdentity, session)) {
+                throw new SemanticException("'admin' user cannot be dropped because of " +
+                        "'authorization_enable_admin_user_protection' configuration is enabled");
+            }
+
+            if (!authenticationManager.doesUserExist(userIdentity) && !stmt.isIfExists()) {
+                throw new SemanticException("Operation DROP USER failed for " + userIdentity + " : user not exists");
+            }
+
+            if (stmt.getUserIdentity().equals(UserIdentity.ROOT)) {
+                throw new SemanticException("Operation DROP USER failed for " + UserIdentity.ROOT +
+                        " : cannot drop user " + UserIdentity.ROOT);
             }
             return null;
         }
 
         @Override
-        public Void visitDropUserStatement(DropUserStmt stmt, ConnectContext session) {
-            analyseUser(stmt.getUserIdentity(), session, false);
+        public Void visitShowAuthenticationStatement(ShowAuthenticationStmt statement, ConnectContext context) {
+            UserIdentity user = statement.getUserIdent();
+            if (user != null) {
+                analyseUser(user, true);
+            } else if (!statement.isAll()) {
+                statement.setUserIdent(context.getCurrentUserIdentity());
+            }
             return null;
         }
 
         @Override
         public Void visitCreateRoleStatement(CreateRoleStmt stmt, ConnectContext session) {
-            validRoleName(stmt.getRoles().get(0), false, "Can not create role");
+            for (String roleName : stmt.getRoles()) {
+                FeNameFormat.checkRoleName(roleName, true, "Can not create role");
+                if (authorizationManager.checkRoleExists(roleName) && !stmt.isIfNotExists()) {
+                    throw new SemanticException("Operation CREATE ROLE failed for " + roleName + " : role already exists");
+                }
+            }
+            return null;
+        }
+
+        @Override
+        public Void visitDropRoleStatement(DropRoleStmt stmt, ConnectContext session) {
+            for (String roleName : stmt.getRoles()) {
+                FeNameFormat.checkRoleName(roleName, true, "Can not create role");
+                if (!authorizationManager.checkRoleExists(roleName) && !stmt.isIfExists()) {
+                    throw new SemanticException("Operation DROP ROLE failed for " + roleName + " : role not exists");
+                }
+            }
+            return null;
+        }
+
+        private ObjectType analyzeObjectType(String objectTypeUnResolved) {
+            String o = objectTypeUnResolved.replace(" ", "_");
+
+            if (!EnumUtils.isValidEnumIgnoreCase(ObjectType.class, o)) {
+                ObjectType objectType = ObjectType.OBJECT_TO_PLURAL.get(objectTypeUnResolved);
+                if (objectType == null) {
+                    throw new SemanticException("cannot find privilege object type " + objectTypeUnResolved);
+                } else {
+                    return objectType;
+                }
+            }
+            return ObjectType.valueOf(o);
+        }
+
+        private PrivilegeType analyzePrivType(ObjectType objectType, String privTypeString) {
+            String p = privTypeString.replace(" ", "_");
+
+            if (!EnumUtils.isValidEnumIgnoreCase(PrivilegeType.class, p)) {
+                throw new SemanticException("cannot find privilege type " + privTypeString);
+            }
+
+            PrivilegeType privilegeType = PrivilegeType.valueOf(p);
+            if (!authorizationManager.isAvailablePirvType(objectType, privilegeType)) {
+                throw new SemanticException("Cannot grant or revoke " + privTypeString + " on '"
+                        + objectType + "' type object");
+            }
+            return privilegeType;
+        }
+
+        @Override
+        public Void visitGrantRevokePrivilegeStatement(BaseGrantRevokePrivilegeStmt stmt, ConnectContext session) {
+            // validate user/role
+            if (stmt.getUserIdentity() != null) {
+                analyseUser(stmt.getUserIdentity(), true);
+            } else {
+                validRoleName(stmt.getRole(), "Can not grant/revoke to role", true);
+            }
+
+            try {
+                ObjectType objectType = analyzeObjectType(stmt.getObjectTypeUnResolved());
+                stmt.setObjectType(objectType);
+
+                List<PEntryObject> objectList = new ArrayList<>();
+                if (stmt.isGrantOnALL()) {
+                    Preconditions.checkArgument(stmt.getPrivilegeObjectNameTokensList() != null);
+                    Preconditions.checkArgument(stmt.getPrivilegeObjectNameTokensList().size() == 1);
+
+                    List<String> tokens = stmt.getPrivilegeObjectNameTokensList().get(0);
+                    switch (objectType) {
+                        case TABLE:
+                            if (tokens.size() != 2) {
+                                throw new SemanticException("Invalid grant statement with error privilege object " + tokens);
+                            }
+                            objectList.add(authorizationManager.generateObject(objectType,
+                                    Lists.newArrayList(session.getCurrentCatalog(), tokens.get(0), tokens.get(1))));
+                            break;
+
+                        case VIEW:
+                        case MATERIALIZED_VIEW:
+                            if (tokens.size() != 2) {
+                                throw new SemanticException("Invalid grant statement with error privilege object " + tokens);
+                            }
+                            objectList.add(authorizationManager.generateObject(objectType, tokens));
+                            break;
+
+                        case DATABASE:
+                            if (tokens.size() != 1) {
+                                throw new SemanticException("Invalid grant statement with error privilege object " + tokens);
+                            }
+                            objectList.add(authorizationManager.generateObject(objectType,
+                                    Lists.newArrayList(session.getCurrentCatalog(), tokens.get(0))));
+                            break;
+
+                        case SYSTEM:
+                            //Not exists grant on all system
+                            break;
+
+                        case USER:
+                            if (tokens.size() != 1) {
+                                throw new SemanticException("Invalid grant statement with error privilege object " + tokens);
+                            }
+                            objectList.add(authorizationManager.generateUserObject(objectType, null));
+                            break;
+
+                        case RESOURCE:
+                        case CATALOG:
+                        case RESOURCE_GROUP:
+                            if (tokens.size() != 1) {
+                                throw new SemanticException("Invalid grant statement with error privilege object " + tokens);
+                            }
+
+                            objectList.add(authorizationManager.generateObject(stmt.getObjectType(), tokens));
+                            break;
+
+                        case FUNCTION:
+                            if (tokens.size() != 2) {
+                                throw new SemanticException("Invalid grant statement with error privilege object " + tokens);
+                            }
+
+                            if (tokens.get(0).equals("*")) {
+                                objectList.add(authorizationManager.generateFunctionObject(objectType,
+                                        PrivilegeBuiltinConstants.ALL_DATABASE_ID,
+                                        PrivilegeBuiltinConstants.ALL_FUNCTIONS_ID));
+                            } else {
+                                Database database = GlobalStateMgr.getServingState().getDb(tokens.get(0));
+                                if (database == null) {
+                                    throw new SemanticException("Database %s is not found", tokens.get(0));
+                                }
+
+                                objectList.add(authorizationManager.generateFunctionObject(objectType,
+                                        database.getId(), PrivilegeBuiltinConstants.ALL_FUNCTIONS_ID));
+                            }
+
+                            break;
+
+                        case GLOBAL_FUNCTION:
+                            if (tokens.size() != 1) {
+                                throw new SemanticException("Invalid grant statement with error privilege object " + tokens);
+                            }
+
+                            objectList.add(authorizationManager.generateFunctionObject(stmt.getObjectType(),
+                                    PrivilegeBuiltinConstants.GLOBAL_FUNCTION_DEFAULT_DATABASE_ID,
+                                    PrivilegeBuiltinConstants.ALL_FUNCTIONS_ID));
+                            break;
+                        default:
+                            throw new SemanticException("Grant/Revoke unsupported object type " + objectType.name());
+                    }
+
+                } else {
+                    switch (objectType) {
+                        case TABLE:
+                            Preconditions.checkArgument(stmt.getPrivilegeObjectNameTokensList() != null);
+
+                            for (List<String> tokens : stmt.getPrivilegeObjectNameTokensList()) {
+                                TableName tableName;
+                                if (tokens.size() == 3) {
+                                    tableName = new TableName(tokens.get(0), tokens.get(1), tokens.get(2));
+                                } else if (tokens.size() == 2) {
+                                    tableName = new TableName(tokens.get(0), tokens.get(1));
+                                    MetaUtils.normalizationTableName(session, tableName);
+                                } else if (tokens.size() == 1) {
+                                    tableName = new TableName("", tokens.get(0));
+                                    MetaUtils.normalizationTableName(session, tableName);
+                                } else {
+                                    throw new SemanticException("Invalid grant statement with error privilege object " + tokens);
+                                }
+
+                                objectList.add(authorizationManager.generateObject(objectType,
+                                        Lists.newArrayList(tableName.getCatalog(), tableName.getDb(), tableName.getTbl())));
+
+                            }
+                            break;
+
+                        case VIEW:
+                        case MATERIALIZED_VIEW:
+                            Preconditions.checkArgument(stmt.getPrivilegeObjectNameTokensList() != null);
+
+                            for (List<String> tokens : stmt.getPrivilegeObjectNameTokensList()) {
+                                TableName tableName;
+                                if (tokens.size() == 2) {
+                                    tableName = new TableName(tokens.get(0), tokens.get(1));
+                                } else if (tokens.size() == 1) {
+                                    tableName = new TableName("", tokens.get(0));
+                                    MetaUtils.normalizationTableName(session, tableName);
+                                } else {
+                                    throw new SemanticException("Invalid grant statement with error privilege object " + tokens);
+                                }
+
+                                objectList.add(authorizationManager.generateObject(objectType,
+                                        Lists.newArrayList(tableName.getDb(), tableName.getTbl())));
+                            }
+                            break;
+
+                        case DATABASE:
+                            Preconditions.checkArgument(stmt.getPrivilegeObjectNameTokensList() != null);
+
+                            for (List<String> tokens : stmt.getPrivilegeObjectNameTokensList()) {
+                                if (tokens.size() == 2) {
+                                    objectList.add(authorizationManager.generateObject(objectType,
+                                            Lists.newArrayList(tokens.get(0), tokens.get(1))));
+                                } else if (tokens.size() == 1) {
+                                    objectList.add(authorizationManager.generateObject(objectType,
+                                            Lists.newArrayList(session.getCurrentCatalog(), tokens.get(0))));
+                                } else {
+                                    throw new SemanticException("Invalid grant statement with error privilege object " + tokens);
+                                }
+                            }
+                            break;
+
+                        case SYSTEM:
+                            objectList.addAll(Arrays.asList(new PEntryObject[] {null}));
+                            break;
+
+                        case USER:
+                            for (UserIdentity userIdentity : stmt.getUserPrivilegeObjectList()) {
+                                analyseUser(userIdentity, true);
+                                objectList.add(authorizationManager.generateUserObject(stmt.getObjectType(), userIdentity));
+                            }
+                            break;
+
+                        case RESOURCE:
+                        case CATALOG:
+                        case RESOURCE_GROUP:
+                            for (List<String> tokens : stmt.getPrivilegeObjectNameTokensList()) {
+                                if (tokens.size() != 1) {
+                                    throw new SemanticException("Invalid grant statement with error privilege object " + tokens);
+                                }
+
+                                objectList.add(authorizationManager.generateObject(stmt.getObjectType(), tokens));
+                            }
+                            break;
+
+                        case FUNCTION:
+                            for (Pair<FunctionName, FunctionArgsDef> f : stmt.getFunctions()) {
+                                FunctionName functionName = f.first;
+                                if (functionName.getDb() == null) {
+                                    String dbName = ConnectContext.get().getDatabase();
+                                    if (dbName.equals("")) {
+                                        throw new SemanticException("database not selected");
+                                    }
+                                    functionName.setDb(dbName);
+                                }
+
+                                FunctionArgsDef argsDef = f.second;
+                                argsDef.analyze();
+                                FunctionSearchDesc searchDesc = new FunctionSearchDesc(functionName,
+                                        argsDef.getArgTypes(), argsDef.isVariadic());
+
+                                Database db = GlobalStateMgr.getCurrentState().getDb(functionName.getDb());
+                                long databaseID = db.getId();
+                                Function function = db.getFunction(searchDesc);
+
+                                if (function == null) {
+                                    throw new SemanticException("cannot find function " + functionName + "!");
+                                } else {
+                                    PEntryObject object = authorizationManager.generateFunctionObject(
+                                            analyzeObjectType(stmt.getObjectTypeUnResolved()), databaseID,
+                                            function.getFunctionId());
+                                    objectList.add(object);
+                                }
+                            }
+                            break;
+
+                        case GLOBAL_FUNCTION:
+                            for (Pair<FunctionName, FunctionArgsDef> f : stmt.getFunctions()) {
+                                FunctionName functionName = f.first;
+                                FunctionArgsDef argsDef = f.second;
+                                argsDef.analyze();
+                                FunctionSearchDesc searchDesc = new FunctionSearchDesc(functionName,
+                                        argsDef.getArgTypes(), argsDef.isVariadic());
+
+                                Function function = GlobalStateMgr.getCurrentState().getGlobalFunctionMgr()
+                                        .getFunction(searchDesc);
+
+                                if (function == null) {
+                                    throw new SemanticException("cannot find function " + functionName + "!");
+                                } else {
+                                    PEntryObject object = authorizationManager.generateFunctionObject(
+                                            analyzeObjectType(stmt.getObjectTypeUnResolved()),
+                                            PrivilegeBuiltinConstants.GLOBAL_FUNCTION_DEFAULT_DATABASE_ID,
+                                            function.getFunctionId());
+                                    objectList.add(object);
+                                }
+                            }
+                            break;
+                        default:
+                            throw new SemanticException("Grant/Revoke unsupported object type " + objectType.name());
+                    }
+                }
+                stmt.setObjectList(objectList);
+
+                List<PrivilegeType> privilegeTypes = new ArrayList<>();
+                for (String privTypeUnResolved : stmt.getPrivilegeTypeUnResolved()) {
+                    if (privTypeUnResolved.equalsIgnoreCase("all")
+                            || privTypeUnResolved.equalsIgnoreCase("all privileges")) {
+                        privilegeTypes.addAll(authorizationManager.getAvailablePrivType(stmt.getObjectType()));
+                    } else {
+                        privilegeTypes.add(analyzePrivType(stmt.getObjectType(), privTypeUnResolved));
+                    }
+                }
+
+                stmt.setPrivilegeTypes(privilegeTypes);
+
+                authorizationManager.validateGrant(stmt.getObjectType(), stmt.getPrivilegeTypes(), stmt.getObjectList());
+            } catch (PrivilegeException | AnalysisException e) {
+                SemanticException exception = new SemanticException(e.getMessage());
+                exception.initCause(e);
+                throw exception;
+            }
+            return null;
+        }
+
+        @Override
+        public Void visitSetRoleStatement(SetRoleStmt stmt, ConnectContext session) {
+            for (String roleName : stmt.getRoles()) {
+                validRoleName(roleName, "Cannot set role", true);
+            }
+            return null;
+        }
+
+        @Override
+        public Void visitSetDefaultRoleStatement(SetDefaultRoleStmt stmt, ConnectContext session) {
+            analyseUser(stmt.getUserIdentity(), true);
+            try {
+                for (String roleName : stmt.getRoles()) {
+                    validRoleName(roleName, "Cannot set role", true);
+
+                    Long roleId = authorizationManager.getRoleIdByNameAllowNull(roleName);
+                    Set<Long> roleIdsForUser = authorizationManager.getRoleIdsByUser(stmt.getUserIdentity());
+                    if (roleId == null || !roleIdsForUser.contains(roleId)) {
+                        throw new SemanticException("Role " + roleName + " is not granted to " +
+                                stmt.getUserIdentity().toString());
+                    }
+                }
+            } catch (PrivilegeException e) {
+                throw new SemanticException(e.getMessage());
+            }
+
+            return null;
+        }
+
+        /**
+         * GRANT rolexx to userxx
+         * GRANT role1 to role role2
+         * REVOKE rolexx from userxx
+         * REVOKE role1 from role role2
+         */
+        @Override
+        public Void visitGrantRevokeRoleStatement(BaseGrantRevokeRoleStmt stmt, ConnectContext session) {
+            if (stmt.getUserIdentity() != null) {
+                analyseUser(stmt.getUserIdentity(), true);
+                if (needProtectAdminUser(stmt.getUserIdentity(), session)) {
+                    throw new SemanticException("roles of 'admin' user cannot be changed because of " +
+                            "'authorization_enable_admin_user_protection' configuration is enabled");
+                }
+                stmt.getGranteeRole().forEach(role ->
+                        validRoleName(role, "Can not granted/revoke role to/from user", true));
+            } else {
+                validRoleName(stmt.getRole(), "Can not granted/revoke role to/from role", true);
+                stmt.getGranteeRole().forEach(role ->
+                        validRoleName(role, "Can not granted/revoke role to/from user", true));
+            }
+            return null;
+        }
+
+        @Override
+        public Void visitExecuteAsStatement(ExecuteAsStmt stmt, ConnectContext session) {
+            if (stmt.isAllowRevert()) {
+                throw new SemanticException("`EXECUTE AS` must use with `WITH NO REVERT` for now!");
+            }
+            analyseUser(stmt.getToUser(), true);
             return null;
         }
 
         @Override
         public Void visitShowGrantsStatement(ShowGrantsStmt stmt, ConnectContext session) {
             if (stmt.getUserIdent() != null) {
-                analyseUser(stmt.getUserIdent(), session, true);
+                analyseUser(stmt.getUserIdent(), true);
+            } else if (stmt.getRole() != null) {
+                validRoleName(stmt.getRole(), "There is no such grant defined for role " + stmt.getRole(), true);
             } else {
                 stmt.setUserIdent(session.getCurrentUserIdentity());
             }
-            Preconditions.checkState(session.getCurrentUserIdentity() != null);
-            return null;
-        }
 
-        @Override
-        public Void visitDropRoleStatement(DropRoleStmt stmt, ConnectContext session) {
-            validRoleName(stmt.getRoles().get(0), false, "Can not drop role");
             return null;
         }
     }
